// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Int or string */
  ActionCount: any;
  /** AreaOfEffectType ("SPHERE", "CUBE", "CYLINDER", "LINE", "CONE") or list of AreaOfEffectTypes */
  AreaOfEffectTypeFilter: any;
  /** Float, list of floats, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  FloatFilter: any;
  /** Int, list of ints, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  IntFilter: any;
  /** LanguageScript ("COMMON", "ELVISH", "DWARVISH", "INFERNAL", "DRACONIC", "CELESTIAL") or list of LanguageScripts */
  LanguageScriptFilter: any;
  /** MonsterSubtype ("ANY_RACE", "HUMAN", "DWARF", "ELF", "GOBLINOID", "MERFOLK", "SHAPECHANGER", "DEMON", "DEVIL", "ORC", "SAHUAGIN", "TITAN", "KOBOLD", "GNOLL", "GRIMLOCK", "LIZARDFOLK", "GNOME") or list of MonsterSubtypes */
  MonsterSubtypeFilter: any;
  /** MonsterTypes ("BEAST", "MONSTROSITY", "DRAGON", "HUMANOID", "UNDEAD", "FIEND", "CELESTIAL", "CONSTRUCT", "GIANT", "ELEMENTAL", "FEY", "ABERRATION", "OOZE", "SWARM", "PLANT") or list of MonsterTypes */
  MonsterTypeFilter: any;
  /** ProficiencyType ("WEAPONS", "ARTISANS_TOOLS", "SKILLS", "ARMOR", "MUSICAL_INSTRUMENTS", "SAVING_THROWS", "OTHER", "GAMING_SETS", "VEHICLES") or list of ProficiencyTypes */
  ProficiencyTypeFilter: any;
  /** Size ("TINY", "SMALL", "MEDIUM", "LARGE", "HUGE", "GARGANTUAN") or list of sizes */
  SizeFilter: any;
  /** SpellAttackType ("MELEE", "RANGED") or list of SpellAttackTypes */
  SpellAttackTypeFilter: any;
  /** String or list of strings */
  StringFilter: any;
};

export type AbilityBonus = {
  __typename?: 'AbilityBonus';
  ability_score: AbilityScore;
  bonus: Scalars['Int'];
};

export type AbilityBonusChoice = {
  __typename?: 'AbilityBonusChoice';
  choose: Scalars['Int'];
  from: AbilityBonusOptionSet;
  type: Scalars['String'];
};

export type AbilityBonusOption = {
  __typename?: 'AbilityBonusOption';
  ability_score: AbilityScore;
  bonus: Scalars['Int'];
  option_type: Scalars['String'];
};

export type AbilityBonusOptionSet = {
  __typename?: 'AbilityBonusOptionSet';
  option_set_type: Scalars['String'];
  options: Array<AbilityBonusOption>;
};

export type AbilityScore = {
  __typename?: 'AbilityScore';
  desc: Array<Scalars['String']>;
  full_name: Scalars['String'];
  index: Scalars['String'];
  name: Scalars['String'];
  skills: Array<Skill>;
};


export type AbilityScoreSkillsArgs = {
  name?: InputMaybe<Scalars['String']>;
  order_direction?: InputMaybe<OrderByDirection>;
};

export type AbilityScorePrerequisite = {
  __typename?: 'AbilityScorePrerequisite';
  ability_score: AbilityScore;
  minimum_score: Scalars['Int'];
};
// ...
export type Ammunition = IEquipment & IEquipmentBase & IGear & {
  __typename?: 'Ammunition';
  cost: Cost;
  desc?: Maybe<Array<Scalars['String']>>;
  equipment_category: EquipmentCategory;
  gear_category: EquipmentCategory;
  index: Scalars['String'];
  name: Scalars['String'];
  quantity: Scalars['Int'];
  weight?: Maybe<Scalars['Float']>;
};

export type AreaOfEffect = {
  __typename?: 'AreaOfEffect';
  size: Scalars['Int'];
  type: AreaOfEffectType;
};

export type AreaOfEffectFilter = {
  size?: InputMaybe<Scalars['IntFilter']>;
  type?: InputMaybe<Scalars['AreaOfEffectTypeFilter']>;
};

export enum AreaOfEffectType {
  Cone = 'CONE',
  Cube = 'CUBE',
  Cylinder = 'CYLINDER',
  Line = 'LINE',
  Sphere = 'SPHERE'
}

export type Armor = IEquipment & IEquipmentBase & {
  __typename?: 'Armor';
  armor_category: EquipmentCategory;
  armor_class: ArmorClass;
  cost: Cost;
  desc?: Maybe<Array<Scalars['String']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String'];
  name: Scalars['String'];
  stealth_disadvantage: Scalars['Boolean'];
  str_minimum: Scalars['Int'];
  weight?: Maybe<Scalars['Float']>;
};
// ...
export type Class = {
  __typename?: 'Class';
  class_levels: Array<Level>;
  hit_die: Scalars['Int'];
  index: Scalars['String'];
  multi_classing: Multiclassing;
  name: Scalars['String'];
  proficiencies: Array<Proficiency>;
  proficiency_choices: Array<ProficiencyChoice>;
  saving_throws: Array<AbilityScore>;
  spellcasting?: Maybe<ClassSpellcasting>;
  spells?: Maybe<Array<Spell>>;
  starting_equipment: Array<Quantity>;
  starting_equipment_options: Array<StartingEquipmentChoice>;
  subclasses: Array<Subclass>;
};
// ...
export type ClassSpecific = BarbarianSpecific | BardSpecific | ClericSpecific | DruidSpecific | FighterSpecific | MonkSpecific | PaladinSpecific | RangerSpecific | RogueSpecific | SorcererSpecific | WarlockSpecific | WizardSpecific;

export type ClassSpellcasting = {
  __typename?: 'ClassSpellcasting';
  info: Array<SpellcastingInfo>;
  level: Scalars['Int'];
  spellcasting_ability: AbilityScore;
};
// ...
export type EquipmentCategory = {
  __typename?: 'EquipmentCategory';
  equipment: Array<IEquipmentBase>;
  index: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type Feature = {
  __typename?: 'Feature';
  class: Class;
  desc: Array<Scalars['String']>;
  feature_specific?: Maybe<FeatureSpecific>;
  index: Scalars['String'];
  level: Scalars['Int'];
  name: Scalars['String'];
  parent?: Maybe<Feature>;
  prerequisites: Array<FeaturePrerequisite>;
  reference?: Maybe<Scalars['String']>;
  subclass?: Maybe<Subclass>;
};
// ...
export type HealingAtLevel = {
  __typename?: 'HealingAtLevel';
  healing: Scalars['String'];
  level: Scalars['Int'];
};

export type IEquipment = {
  cost: Cost;
  desc?: Maybe<Array<Scalars['String']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String'];
  name: Scalars['String'];
  weight?: Maybe<Scalars['Float']>;
};
// ...
export type Language = {
  __typename?: 'Language';
  desc?: Maybe<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
  script?: Maybe<LanguageScript>;
  type: LanguageType;
  typical_speakers: Array<Scalars['String']>;
};

export type LanguageChoice = {
  __typename?: 'LanguageChoice';
  choose: Scalars['Int'];
  from: LanguageOptionSet;
  type: Scalars['String'];
};
// ...
export type Level = {
  __typename?: 'Level';
  ability_score_bonuses?: Maybe<Scalars['Int']>;
  class: Class;
  class_specific?: Maybe<ClassSpecific>;
  features: Array<Feature>;
  index: Scalars['String'];
  level: Scalars['Int'];
  prof_bonus?: Maybe<Scalars['Int']>;
  spellcasting?: Maybe<LevelSpellcasting>;
  subclass?: Maybe<Subclass>;
  subclass_specific?: Maybe<SubclassSpecific>;
};
// ...
export type LevelSpellcasting = {
  __typename?: 'LevelSpellcasting';
  cantrips_known?: Maybe<Scalars['Int']>;
  spell_slots_level_1?: Maybe<Scalars['Int']>;
  spell_slots_level_2?: Maybe<Scalars['Int']>;
  spell_slots_level_3?: Maybe<Scalars['Int']>;
  spell_slots_level_4?: Maybe<Scalars['Int']>;
  spell_slots_level_5?: Maybe<Scalars['Int']>;
  spell_slots_level_6?: Maybe<Scalars['Int']>;
  spell_slots_level_7?: Maybe<Scalars['Int']>;
  spell_slots_level_8?: Maybe<Scalars['Int']>;
  spell_slots_level_9?: Maybe<Scalars['Int']>;
  spells_known?: Maybe<Scalars['Int']>;
};
// ...
export type MagicSchool = {
  __typename?: 'MagicSchool';
  desc: Scalars['String'];
  index: Scalars['String'];
  name: Scalars['String'];
  spells: Array<Spell>;
};
// ...
export type Multiclassing = {
  __typename?: 'Multiclassing';
  prerequisite_options?: Maybe<PrerequisiteChoice>;
  prerequisites?: Maybe<Array<AbilityScorePrerequisite>>;
  proficiencies: Array<Proficiency>;
  proficiency_choices?: Maybe<Array<ProficiencyChoice>>;
};
// ...
export type PrerequisiteChoice = {
  __typename?: 'PrerequisiteChoice';
  choose: Scalars['Int'];
  from: PrerequisiteOptionSet;
  type: Scalars['String'];
};

export type PrerequisiteOption = {
  __typename?: 'PrerequisiteOption';
  ability_score: AbilityScore;
  minimum_score: Scalars['Int'];
  option_type: Scalars['String'];
};

export type PrerequisiteOptionSet = {
  __typename?: 'PrerequisiteOptionSet';
  option_set_type: Scalars['String'];
  options: Array<PrerequisiteOption>;
};

export type Proficiency = {
  __typename?: 'Proficiency';
  classes: Array<Class>;
  index: Scalars['String'];
  name: Scalars['String'];
  races: Array<ProficiencyRace>;
  reference: ProficiencyReference;
  type: ProficiencyType;
};


export type ProficiencyClassesArgs = {
  name?: InputMaybe<Scalars['String']>;
};


export type ProficiencyRacesArgs = {
  name?: InputMaybe<Scalars['String']>;
};

export type ProficiencyChoice = {
  __typename?: 'ProficiencyChoice';
  choose: Scalars['Int'];
  desc?: Maybe<Scalars['String']>;
  from: ProficiencyOptionSet;
  type: Scalars['String'];
};

export type ProficiencyChoiceOption = {
  __typename?: 'ProficiencyChoiceOption';
  choice: ProficiencyChoice;
  option_type: Scalars['String'];
};

export type ProficiencyOption = ProficiencyChoiceOption | ProficiencyReferenceOption;

export type ProficiencyOptionSet = {
  __typename?: 'ProficiencyOptionSet';
  option_set_type: Scalars['String'];
  options: Array<ProficiencyOption>;
};

export type ProficiencyOrder = {
  by: ProficiencyOrderBy;
  direction?: OrderByDirection;
  then_by?: InputMaybe<ProficiencyOrder>;
};

export enum ProficiencyOrderBy {
  Name = 'NAME',
  Type = 'TYPE'
}

export type ProficiencyPrerequisite = {
  __typename?: 'ProficiencyPrerequisite';
  proficiency: Proficiency;
  type: Scalars['String'];
};

export type ProficiencyRace = {
  ability_bonuses: Array<AbilityBonus>;
  index: Scalars['String'];
  name: Scalars['String'];
};

export type ProficiencyReference = AbilityScore | Ammunition | Armor | EquipmentCategory | Gear | Pack | Skill | Tool | Vehicle | Weapon;

export type ProficiencyReferenceOption = {
  __typename?: 'ProficiencyReferenceOption';
  item: Proficiency;
  option_type: Scalars['String'];
};

export enum ProficiencyType {
  Armor = 'ARMOR',
  ArtisansTools = 'ARTISANS_TOOLS',
  GamingSets = 'GAMING_SETS',
  MusicalInstruments = 'MUSICAL_INSTRUMENTS',
  Other = 'OTHER',
  SavingThrows = 'SAVING_THROWS',
  Skills = 'SKILLS',
  Vehicles = 'VEHICLES',
  Weapons = 'WEAPONS'
}

export type Quantity = {
  __typename?: 'Quantity';
  equipment: IEquipment;
  quantity: Scalars['Int'];
};
// ...
export type Race = ProficiencyRace & {
  __typename?: 'Race';
  ability_bonus_options?: Maybe<AbilityBonusChoice>;
  ability_bonuses: Array<AbilityBonus>;
  age: Scalars['String'];
  alignment: Scalars['String'];
  index: Scalars['String'];
  language_desc: Scalars['String'];
  language_options?: Maybe<LanguageChoice>;
  languages: Array<Language>;
  name: Scalars['String'];
  size: Size;
  size_description: Scalars['String'];
  speed: Scalars['Int'];
  starting_proficiencies: Array<Proficiency>;
  starting_proficiency_options?: Maybe<ProficiencyChoice>;
  subraces: Array<Subrace>;
  traits: Array<Trait>;
};
// ...
export enum Size {
  Gargantuan = 'GARGANTUAN',
  Huge = 'HUGE',
  Large = 'LARGE',
  Medium = 'MEDIUM',
  Small = 'SMALL',
  Tiny = 'TINY'
}

export type Skill = {
  __typename?: 'Skill';
  ability_score: AbilityScore;
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type Spell = {
  __typename?: 'Spell';
  area_of_effect?: Maybe<AreaOfEffect>;
  attack_type?: Maybe<SpellAttackType>;
  casting_time: Scalars['String'];
  classes: Array<Class>;
  components?: Maybe<Array<Maybe<SpellComponent>>>;
  concentration: Scalars['Boolean'];
  damage?: Maybe<SpellDamage>;
  dc?: Maybe<SpellDc>;
  desc: Array<Scalars['String']>;
  duration: Scalars['String'];
  heal_at_slot_level?: Maybe<Array<HealingAtLevel>>;
  higher_level?: Maybe<Array<Scalars['String']>>;
  index: Scalars['String'];
  level: Scalars['Int'];
  material?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  range: Scalars['String'];
  ritual: Scalars['Boolean'];
  school: MagicSchool;
  subclasses: Array<Subclass>;
};


export type SpellClassesArgs = {
  name?: InputMaybe<Scalars['String']>;
};


export type SpellSubclassesArgs = {
  name?: InputMaybe<Scalars['String']>;
};

export enum SpellAttackType {
  Melee = 'MELEE',
  Ranged = 'RANGED'
}

export type SpellChoice = {
  __typename?: 'SpellChoice';
  choose: Scalars['Int'];
  from: SpellOptionSet;
  type: Scalars['String'];
};

export enum SpellComponent {
  M = 'M',
  S = 'S',
  V = 'V'
}

export type SpellDamage = {
  __typename?: 'SpellDamage';
  damage_at_character_level?: Maybe<Array<DamageAtLevel>>;
  damage_at_slot_level?: Maybe<Array<DamageAtLevel>>;
  damage_type?: Maybe<DamageType>;
};

export type SpellDc = {
  __typename?: 'SpellDc';
  desc?: Maybe<Scalars['String']>;
  success: DcSuccess;
  type: AbilityScore;
};
// ...
export type SpellWithPrerequisite = {
  __typename?: 'SpellWithPrerequisite';
  prerequisites: Array<Maybe<SpellPrerequisite>>;
  spell: Spell;
};

export type SpellcastingInfo = {
  __typename?: 'SpellcastingInfo';
  desc: Array<Scalars['String']>;
  name: Scalars['String'];
};

export type StartingEquipmentChoice = {
  __typename?: 'StartingEquipmentChoice';
  choose: Scalars['Int'];
  desc: Scalars['String'];
  from: StartingEquipmentOptionSet;
  type: Scalars['String'];
};

export type StartingEquipmentOptionSet = EquipmentCategoryOptionSet | EquipmentOptionSet;
// ...
export type Subclass = {
  __typename?: 'Subclass';
  class: Class;
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
  spells?: Maybe<Array<SpellWithPrerequisite>>;
  subclass_flavor: Scalars['String'];
  subclass_levels: Array<Maybe<Level>>;
};
// ...
export type SubclassSpecific = DevotionSpecific | LoreSpecific;

export type Subrace = ProficiencyRace & {
  __typename?: 'Subrace';
  ability_bonuses: Array<AbilityBonus>;
  desc: Scalars['String'];
  index: Scalars['String'];
  language_options?: Maybe<LanguageChoice>;
  name: Scalars['String'];
  race: Race;
  racial_traits: Array<Trait>;
  starting_proficiencies: Array<Proficiency>;
};
// ...
export type Trait = {
  __typename?: 'Trait';
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  language_options?: Maybe<LanguageChoice>;
  name: Scalars['String'];
  parent?: Maybe<Trait>;
  proficiencies: Array<Proficiency>;
  proficiency_choices?: Maybe<ProficiencyChoice>;
  races: Array<Maybe<Race>>;
  subraces: Array<Subrace>;
  trait_specific?: Maybe<TraitSpecific>;
};